# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface SendOpenrouterMessageBody {
  content: string;
  /** OpenRouter model ID e.g. moonshotai/kimi-k2.6 */
  model: string;
}

export type TwilioPhoneNumberCapabilities = {
  voice: boolean;
  sms: boolean;
  mms: boolean;
};

export interface TwilioPhoneNumber {
  sid: string;
  phoneNumber: string;
  friendlyName: string;
  capabilities: TwilioPhoneNumberCapabilities;
}

export interface SendSmsBody {
  to: string;
  from: string;
  body: string;
}

export interface SendSmsResponse {
  sid: string;
  status: string;
  to: string;
  from: string;
  body: string;
}

export interface MakeCallBody {
  to: string;
  from: string;
  twiml: string;
}

export interface MakeCallResponse {
  sid: string;
  status: string;
  to: string;
  from: string;
}

export interface VoiceTokenBody {
  identity: string;
}

export interface VoiceTokenResponse {
  token: string;
  identity: string;
  ttl: number;
}

export interface LookupBody {
  phoneNumber: string;
  fields?: string[];
}

export type LookupResponseLineTypeIntelligence = {
  [key: string]: unknown;
} | null;

export type LookupResponseCallerName = { [key: string]: unknown } | null;

export interface LookupResponse {
  phoneNumber: string;
  nationalFormat: string;
  countryCode: string;
  valid: boolean;
  lineTypeIntelligence?: LookupResponseLineTypeIntelligence;
  callerName?: LookupResponseCallerName;
}

export interface TwilioAccountInfo {
  sid: string;
  friendlyName: string;
  status: string;
  balance: string;
  currency: string;
}