# lib/api-zod/src/generated/types/lookupResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LookupResponseCallerName } from "./lookupResponseCallerName";
import type { LookupResponseLineTypeIntelligence } from "./lookupResponseLineTypeIntelligence";

export interface LookupResponse {
  phoneNumber: string;
  nationalFormat: string;
  countryCode: string;
  valid: boolean;
  lineTypeIntelligence?: LookupResponseLineTypeIntelligence;
  callerName?: LookupResponseCallerName;
}