# lib/api-zod/src/generated/types/twilioPhoneNumber.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TwilioPhoneNumberCapabilities } from "./twilioPhoneNumberCapabilities";

export interface TwilioPhoneNumber {
  sid: string;
  phoneNumber: string;
  friendlyName: string;
  capabilities: TwilioPhoneNumberCapabilities;
}